Corrigindo prisão aleatória de viewers.
