+++ .env.example
@@ -1,5 +1,4 @@
 BOT_USERNAME=nomedobot
-
 OAUTH_TOKEN=tokenaqui
-
 CHANNEL_NAME=nomedocanal
+MINUTOS_ENTRE_PRISAO_DE_VIEWERS=30
